Split format from export logic
